refactor: Rename markdown.ts to html_render.ts
- Rename src/utils/markdown.ts to src/utils/html_render.ts
- Update all import references
- Expand abbreviated variable names for clarity:
  - vw/vh → viewportWidth/viewportHeight
  - de → docElement
  - sw/sh → scrollWidth/scrollHeight
  - el → element

The file primarily handles HTML-to-image rendering with Puppeteer,
so html_render.ts better reflects its purpose.
