+++ src/actions/captions.ts
@@ -4,7 +4,7 @@ import type { GraphData } from "graphai";
 import * as agents from "@graphai/vanilla";
 import { getHTMLFile, getCaptionImagePath, getOutputStudioFilePath } from "../utils/file.js";
 import { localizedText, processLineBreaks } from "../utils/utils.js";
-import { renderHTMLToImage, interpolate } from "../utils/markdown.js";
+import { renderHTMLToImage, interpolate } from "../utils/html_render.js";
 import { MulmoStudioContextMethods, MulmoPresentationStyleMethods } from "../methods/index.js";
 import { fileWriteAgent } from "@graphai/vanilla_node_agents";
